Revert "Save and restore state of checkboxes"
This reverts commit 2896a6efa3b16634be56df99b14a69d05071cbb9.

Without this patch if a user checks a component and hits 'apply' and then
checks another component and hits 'apply', the latter component will go unchecked.

At the time checkbox state was not being saved and the reverted commit fixed that problem however
this no longer seems to be necessary.

Change-Id: I70383c0d9b6a1d7c70b25d7607d1e2ddc29b2a7f